requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pycountry==22.3.5
requests==2.31.0

